% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_n.R
\name{repeat_n}
\alias{repeat_n}
\title{Replication for randomization and resampling}
\usage{
repeat_n(expr, n = 1, cull = "default", combine = "default",
  names = NULL, ...)
}
\arguments{
\item{expr}{and expression to be repeated}

\item{n}{an integer number of times to repeat}

\item{cull}{a function for culling the results of each evaluation of expr}

\item{names}{names for columns of the returned object}

\item{...}{additional arguments (currently ignored)}
}
\value{
a list or data frame containing the (culled) restuls of the
  repetition
}
\description{
Replication for randomization and resampling
}
